feat: complete P1+P2 roadmap with 22 parallel agents
## P1 Features (4/4)
- Theme toggle (light/dark/system) with FOUC prevention
- Keyboard shortcuts (R/T/C/S/M/?) with help modal
- GenAI tool definitions viewer (tools tab, schemas, calls)
- MCP API key authentication (x-mcp-api-key header)

## P2 Features (4/4)
- Download logs (JSON/CSV export on all pages)
- Health check UI (status indicator + detail modal)
- Text visualizer (JSON/XML auto-detect, syntax highlight)
- Pause/Clear telemetry (controls + DELETE endpoint)

## Infrastructure
- HTTP Protobuf OTLP endpoint (application/x-protobuf)
- DuckDB critical fix: write ops were using read connections
- SQL injection fix in delete methods (parameterized queries)
- SQL statement caching for batch inserts

## Quality
- OTel semconv 1.39 audit (95/100 compliance)
- Roslyn analyzers: 2 new code fixes (QYL013, QYL014)
- Test coverage: 67 analyzer tests
- React 19 migration (ElementRef → ComponentRef)
- SpanRecord type alignment (int→long, decimal→double)

## MCP Server
- 12 new tools (Aspire parity)
- Console, structured log, GenAI, storage tools
- Improved tool descriptions for AI agents

## ServiceDefaults Generator
- Fixed Azure.AI.OpenAI provider ordering
- Fixed Azure token property names
- Added image_generation, speech, transcription, rerank ops
- Thread-safe TracedInterceptorEmitter

Note: CodeFixResources.Designer.cs suppression is auto-generated by VS.

+++ src/qyl.protocol/Models/SpanRecord.cs
@@ -12,8 +12,8 @@ using Qyl.OTel.Enums;
 namespace Qyl.Models;
 
 /// <summary>
-/// Span record for storage and query operations.
-/// Contains flattened fields for efficient database storage.
+///     Span record for storage and query operations.
+///     Contains flattened fields for efficient database storage.
 /// </summary>
 public sealed record SpanRecord
 {
@@ -81,11 +81,11 @@ public sealed record SpanRecord
 
     /// <summary>Input token count</summary>
     [JsonPropertyName("gen_ai_input_tokens")]
-    public int? GenAiInputTokens { get; init; }
+    public long? GenAiInputTokens { get; init; }
 
     /// <summary>Output token count</summary>
     [JsonPropertyName("gen_ai_output_tokens")]
-    public int? GenAiOutputTokens { get; init; }
+    public long? GenAiOutputTokens { get; init; }
 
     /// <summary>Temperature setting</summary>
     [JsonPropertyName("gen_ai_temperature")]
@@ -105,7 +105,7 @@ public sealed record SpanRecord
 
     /// <summary>Cost in USD</summary>
     [JsonPropertyName("gen_ai_cost_usd")]
-    public decimal? GenAiCostUsd { get; init; }
+    public double? GenAiCostUsd { get; init; }
 
     // === JSON Blobs ===
